requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
lightgbm==4.4.0
imbalanced-learn==0.12.3     # SMOTE para desbalanceamento

# Otimização de hiperparâmetros
optuna==3.6.1
optuna-integration==3.6.0

# Explicabilidade
shap==0.45.1

# Rastreamento de experimentos
mlflow==2.14.1

# Validação de dados
pydantic==2.8.2
pydantic-settings==2.3.4

# API
fastapi==0.111.1
uvicorn[standard]==0.30.1
httpx==0.27.0                # client HTTP async para Groq API

# LLM
ollama==0.3.0                # client Ollama local

# Interface
streamlit==1.36.0
plotly==5.22.0

# Serialização
joblib==1.4.2

# Utilitários
python-dotenv==1.0.1
pyyaml==6.0.1
loguru==0.7.2                # logging estruturado
